Enum grid::Dihedral

pub enum Dihedral {
    R0,
    R1,
    R2,
    R3,
    R0S,
    R1S,
    R2S,
    R3S,
}

An element of a dihedral group D₄.

Formally, D₄ is generated by two element R and S subjecting the following relations,

R⁴ = 1,

S² = 1,

SRS = R³.

Each variant of the enum Dihedral corresponds an element of D₄ as this.

R0 : 1

R1 : R

R2 : R²

R3 : R³

R0S : S

R1S : RS

R2S : R²S

R3S : R³S

Variants

R0

R1

R2

R3

R0S

R1S

R2S

R3S