Struct grid::Grid

pub struct Grid<T> { /* private fields */ }

A wrapper to rotate/reflect a grid logically.

Use Dihedral to act this. An affine representation of Dihedral on Map(ℤ², T) is defined by

R1 is a 90-degrees rotation, that is (R1 * f)(i, j) = f(w - 1 - j, h).

R0S is a transposition, that is (R0S * f)(i, j) = f(j, i).

Thus each the element acts like this:

R0: identity.

R1: rotate left.

R2: anti-position.

R3: rotate right.

R0S: transposition.

R1S: reflection through i = (h - 1) / 2.

R2S: anti-transposition.

R3S: reflections through j = (w - 1) / 2.

Implementations

impl<T> Grid<T>

pub fn new(elm: Vec<T>, h: usize, w: usize) -> Self

Creates a new Grid.

Example

Basic usage:

use grid::Grid;

let grid = Grid::new(vec![0, 1, 2, 3], 2, 2);
assert_eq!(grid.h(), 2);

pub fn surround(self, default: T) -> Self

Returns a Grid object surround by infinite default logically.

Example

Basic usage:

use grid::Grid;

let grid = Grid::new(vec![0, 1, 2, 3], 2, 2).surround(std::u32::MAX);
assert_eq!(grid.iget_or_default(0, 0), 0);
assert_eq!(grid.iget_or_default(1, 0), 2);
assert_eq!(grid.iget_or_default(-1, 0), std::u32::MAX);

pub fn h(&self) -> usize

Returns a logical height of a grid.

Example

Basic usage:

use grid::{Grid, Dihedral};

let mut grid = Grid::new(vec![0, 1], 1, 2);
assert_eq!(grid.h(), 1);

grid.apply(Dihedral::R0S);
assert_eq!(grid.h(), 2);

pub fn w(&self) -> usize

Returns a logical width of a grid.

Example

Basic usage:

use grid::{Grid, Dihedral};

let mut grid = Grid::new(vec![0, 1], 1, 2);
assert_eq!(grid.w(), 2);

grid.apply(Dihedral::R0S);
assert_eq!(grid.w(), 1);

pub fn iget_or_default(&self, i: isize, j: isize) -> Twhere T: Copy,

If (i, j) is within the domain, returns the corresponding value, while otherwise returns the “default value” specified via Grid::surround.

Example

Basic usage:

use grid::Grid;

let mut grid = Grid::new(vec![0, 1], 1, 2).surround(-1);
assert_eq!(grid.iget_or_default(0, 0), 0);
assert_eq!(grid.iget_or_default(9, 0), -1);

pub fn iget(&self, i: isize, j: isize) -> Option<&T>

If (i, j) is within the domain, returns a reference to the corresponding value, while otherwise returns None.

Example

Basic usage:

use grid::Grid;

let mut grid = Grid::new(vec![0, 1], 1, 2).surround(-1);
assert_eq!(grid.iget(0, 0), Some(&0));
assert_eq!(grid.iget(9, 0), None);

pub fn iget_mut(&mut self, i: isize, j: isize) -> Option<&mut T>

If (i, j) is within the domain, returns a mutable reference to the corresponding value, while otherwise returns None.

Example

Basic usage:

use grid::Grid;

let mut grid = Grid::new(vec![0, 1], 1, 2).surround(-1);
assert_eq!(grid[(0, 0)], 0);
*grid.iget_mut(0, 0).unwrap() = 3;
assert_eq!(grid[(0, 0)], 3);

pub fn get(&self, i: usize, j: usize) -> Option<&T>

If (i, j) is within the domain, returns a reference to the corresponding value, while otherwise returns None.

Example

Basic usage:

use grid::Grid;

let mut grid = Grid::new(vec![0, 1], 1, 2).surround(-1);
assert_eq!(grid.get(0, 0), Some(&0));
assert_eq!(grid.get(9, 0), None);

pub fn get_mut(&mut self, i: isize, j: isize) -> Option<&mut T>

If (i, j) is within the domain, returns a mutable reference to the corresponding value, while otherwise returns None.

Example

Basic usage:

use grid::Grid;

let mut grid = Grid::new(vec![0, 1], 1, 2).surround(-1);
assert_eq!(grid[(0, 0)], 0);
*grid.get_mut(0, 0).unwrap() = 3;
assert_eq!(grid[(0, 0)], 3);

pub fn apply(&mut self, d: Dihedral)

Apply a dihedral element to a grid logically.

Example

Basic usage:

use grid::{Grid, Dihedral};

let mut grid = Grid::new(vec![0, 1], 1, 2);
assert_eq!(grid.get(0, 1), Some(&1));
assert_eq!(grid.get(1, 0), None);
grid.apply(Dihedral::R0S);
assert_eq!(grid.get(0, 1), None);
assert_eq!(grid.get(1, 0), Some(&1));

pub fn applied(self, d: Dihedral) -> Self

Returns a grid applied a dihedral element logically.

Example

Basic usage:

use grid::{Grid, Dihedral};

let grid = Grid::new(vec![0, 1], 1, 2);
assert_eq!(grid.get(0, 1), Some(&1));
assert_eq!(grid.get(1, 0), None);
let grid = grid.applied(Dihedral::R0S);
assert_eq!(grid.get(0, 1), None);
assert_eq!(grid.get(1, 0), Some(&1));

pub fn rows(&self) -> Rows<'_, T>

Returns an iterator over the logical rows of this grid.

Examples

Basic usage:

use grid::Grid;

let grid = Grid::new("01234567".chars().collect::<Vec<_>>(), 2, 4);
let mut rows = grid.rows();
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['0', '1', '2', '3']);
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['4', '5', '6', '7']);
assert_eq!(rows.next(), None);

pub fn cells(&self) -> impl Iterator<Item = &T>

Returns an iterator over the logical rows of this grid.

Examples

Basic usage:

use grid::Grid;

let grid = Grid::new("01234567".chars().collect::<Vec<_>>(), 2, 4);
let mut rows = grid.rows();
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['0', '1', '2', '3']);
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['4', '5', '6', '7']);
assert_eq!(rows.next(), None);

pub fn collect_vec2(&self) -> Vec<Vec<T>>where T: Clone,

Converts into a 2d vector.

Examples

Basic usage:

use grid::Grid;

let grid = Grid::new("01234567".chars().collect::<Vec<_>>(), 2, 4);
let mut rows = grid.rows();
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['0', '1', '2', '3']);
assert_eq!(rows.next().unwrap().iter().copied().collect::<Vec<_>>(), vec!['4', '5', '6', '7']);
assert_eq!(rows.next(), None);